backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
drf-yasg==1.21.7
python-dotenv==1.0.1
mysqlclient==2.2.4
gunicorn==21.2.0
Pillow==10.2.0
